@@ -31,6 +31,7 @@
       }
     },
     "preset": "ts-jest",
+    "resolver": "./jest-resolver.js",
     "roots": [
       "<rootDir>/packages/babel-preset",
       "<rootDir>/packages/eslint-plugin",
@@ -48,14 +49,15 @@
     "@babel/cli": "7.14.5",
     "@babel/core": "7.14.6",
     "@babel/runtime": "7.14.6",
-    "@types/eslint": "7.29.0",
+    "@types/eslint": "^8.2.1",
     "@types/find-cache-dir": "3.2.1",
     "@types/jest": "27.0.3",
     "@types/listr": "0.14.4",
     "@types/lodash": "4.14.177",
     "@types/node": "16.11.7",
     "@types/prettier": "2.4.1",
-    "eslint": "7.32.0",
+    "enhanced-resolve": "^5.8.3",
+    "eslint": "^8.5.0",
     "husky": "7.0.4",
     "jest": "27.3.1",
     "jest-watch-typeahead": "1.0.0",
