@@ -1,5 +1,5 @@
 import { run } from '@oclif/command';
+import * as flush from '@oclif/command/flush';
 import { handle } from '@oclif/errors';
-import flush = require('@oclif/command/flush');
 
 run().then(flush, handle);
